Vacation Booking Portal
This project implements a simple vacation booking portal with 2PC. The
portal is the coordinator and it interacts with 2 participants, namely,
hotel reservation and concert reservation. The coordinator and the
participants are referred to as entities. Implementation also has a
provision to make the coordinator and the participants artificially fail
and recover.
The hotel will have limited number of rooms and the concert has limited
number of tickets for each day. A vacation can be successfully booked
only if a room is available for all of the days and concert tickets are
available for all the days. Even if one of them is not available the
booking should not be performed. The system should work correctly even
in the face of failures of entities. Program implements various
primitives of the 2PC protocol and use them for implementing the system.